Renaming subcomponents
